backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
pymongo
httpx
langchain
langgraph
sentence-transformers
torch
scikit-learn
rapidfuzz
groq
python-multipart